requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
aiofiles>=0.7.0
Jinja2>=3.0.2,<3.0.3
uvicorn[standard]>=0.15,<0.16
requests>=2.26.0
websockets>=10.0.0,<10.0.1
python-multipart>=0.0.5
Pillow>=8.4.0
sqlalchemy>=1.4.32,<1.4.33
databases>=0.5.5,<0.5.6
aiosqlite>=0.17.0,<0.17.1
pydantic>=1.9.0,<1.9.1
python-multipart>=0.0.5<0.0.6

